Add the python dependencies
As specified in the requirements.txt and setup.py.

They are needed by plpy procedural language, which uses the python
system installation and python modules that must be installed at system
level with their c/c++ bindings.

We have some experience with those: getting a working installation is
easy, but getting it right is not obvious at all.

These are the tested and packages versions, so for the moment I'm
pinning to those.